Removed unused Param class and moved security -> auth
