+++ src/main/java/org/radarcns/auth/NeedsPermissionOnProject.java
@@ -1,7 +1,6 @@
-package org.radarcns.security.filter;
+package org.radarcns.auth;
 
 import static org.radarcns.webapp.util.Parameter.PROJECT_NAME;
-import static org.radarcns.webapp.util.Parameter.SUBJECT_ID;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,9 +8,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import org.radarcns.auth.authorization.Permission;
 
+/**
+ * Indicates that a method needs an authenticated user that has a certain permission on a project.
+ */
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface NeedsPermissionOnSubject {
+public @interface NeedsPermissionOnProject {
     /**
      * Entity that the permission is needed on.
      */
@@ -24,15 +26,8 @@ public @interface NeedsPermissionOnSubject {
 
     /**
      * Path parameter for the name of the project that the entity belongs to.
-     * In this API, {@link org.radarcns.webapp.util.Parameter#PROJECT_NAME} is usually used for path
-     * parameters, so that is the default.
+     * In this API, this is usually {@link org.radarcns.webapp.util.Parameter#PROJECT_NAME}, so
+     * this is the default.
      */
     String projectParam() default PROJECT_NAME;
-
-    /**
-     * Path parameter for the ID of the subject that the entity belongs to.
-     * In this API, {@link org.radarcns.webapp.util.Parameter#SUBJECT_ID} is usually used for path
-     * parameters, so that is the default.
-     */
-    String subjectParam() default SUBJECT_ID;
 }
