+++ src/main/java/org/radarcns/auth/PermissionOnProjectFilter.java
@@ -1,6 +1,6 @@
-package org.radarcns.security.filter;
+package org.radarcns.auth;
 
-import static org.radarcns.security.filter.PermissionFilter.abortWithForbidden;
+import static org.radarcns.auth.PermissionFilter.abortWithForbidden;
 
 import java.io.IOException;
 import javax.ws.rs.container.ContainerRequestContext;
