+++ src/main/java/org/radarcns/webapp/resource/ProjectEndPoint.java
@@ -19,11 +19,11 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.ext.Provider;
+import org.radarcns.auth.NeedsPermission;
+import org.radarcns.auth.NeedsPermissionOnProject;
 import org.radarcns.listener.managementportal.ManagementPortalClient;
 import org.radarcns.managementportal.Project;
 import org.radarcns.managementportal.Subject;
-import org.radarcns.security.filter.NeedsPermission;
-import org.radarcns.security.filter.NeedsPermissionOnProject;
 
 @Provider
 @Path("/" + PROJECTS)
