+++ src/main/java/org/radarcns/webapp/resource/SubjectEndPoint.java
@@ -37,14 +37,14 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.ext.Provider;
+import org.radarcns.auth.NeedsPermissionOnProject;
+import org.radarcns.auth.NeedsPermissionOnSubject;
 import org.radarcns.auth.authorization.Permission.Entity;
 import org.radarcns.dao.SubjectDataAccessObject;
 import org.radarcns.listener.managementportal.ManagementPortalClient;
 import org.radarcns.restapi.header.EffectiveTimeFrame;
 import org.radarcns.restapi.subject.Cohort;
 import org.radarcns.restapi.subject.Subject;
-import org.radarcns.security.filter.NeedsPermissionOnProject;
-import org.radarcns.security.filter.NeedsPermissionOnSubject;
 import org.radarcns.webapp.validation.Alphanumeric;
 
 /**
